Safari-611.1.10/LayoutTests/pointer-lock/lock-lost-on-esc-in-fullscreen-expected.txt

Test that pointerlockchange event is dispatched on esc in fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.testRunner is defined.
PASS onwebkitfullscreenchange received after: Entered fullscreen
     Lock targetDiv1.
PASS onpointerlockchange received after: Lock targetDiv1.
PASS document.pointerLockElement is targetDiv1
Sending Escape keyDown event.
PASS onpointerlockchange received after: Sending Escape keyDown event.
PASS onwebkitfullscreenchange received after: Sending Escape keyDown event.
PASS document.pointerLockElement is null
PASS document.webkitIsFullScreen is false
PASS successfullyParsed is true

TEST COMPLETE
